fix (dev_spring_boot) renamed category service classes
