--- src/main/java/com/project/category/service/CrudCategoryService.java
@@ -1,18 +1,0 @@
-package com.project.category.service;
-
-import com.github.fge.jsonpatch.mergepatch.JsonMergePatch;
-import com.project.category.model.CategoryDTO;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface CrudCategoryService {
-    CategoryDTO create(CategoryDTO categoryDto);
-
-    Page<CategoryDTO> getAll(Pageable pageable);
-
-    CategoryDTO get(Long id);
-
-    CategoryDTO update(Long id, JsonMergePatch patch);
-
-    void delete(Long id);
-}
